requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.9.1
bleach==6.0.0
Brotli==1.0.9
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.6
cmake==3.26.1
codecarbon==2.3.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
coverage==7.3.2
cycler==0.11.0
DataProperty==1.0.1
datasets==2.14.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.19.3
dill==0.3.6
distlib==0.3.7
evaluate==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.4
flake8==6.1.0
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
fuzzywuzzy==0.18.0
huggingface-hub==0.15.1
hydra-core==1.3.2
identify==2.5.30
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
inflate64==0.3.1
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.3
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.25.0
kiwisolver==1.4.4
lit==16.0.0
-e git+https://github.com/sashavor/lm-evaluation-harness@8fbf02529d6124156ce9bcf873e14f35576aac10#egg=lm_eval
lxml==4.9.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mccabe==0.7.0
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.2
numexpr==2.8.7
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pathvalidate==3.2.0
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.11.0
plotly==5.16.1
pluggy==1.3.0
portalocker==2.8.2
pre-commit==3.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.20.6
pyarrow==11.0.0
pybcj==1.0.1
pybind11==2.11.1
pycodestyle==2.11.0
pycparser==2.21
pycryptodomex==3.18.0
pyflakes==3.1.0
Pygments==2.15.1
pyJoules==0.5.1
pynvml==11.5.0
pyparsing==3.0.9
pyppmd==1.0.0
pyrsistent==0.19.3
pytablewriter==1.1.0
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
pyzstd==0.15.9
qtconsole==5.4.2
QtPy==2.3.1
referencing==0.30.2
regex==2023.3.23
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge-score==0.1.2
rpds-py==0.10.6
sacrebleu==2.3.1
safetensors==0.3.1
scikit-learn==1.3.1
scipy==1.11.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlitedict==2.1.0
stack-data==0.6.2
sympy==1.11.1
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tenacity==8.2.3
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.2.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3
tqdm==4.65.0
tqdm-multiprocess==0.0.11
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
typepy==1.3.2
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
virtualenv==20.24.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
zstandard==0.21.0
